// src/app/Model/register.ts
export class Register {
    Title: string = '';
    FirstName: string = '';
    LastName: string = '';
    PhoneNumber: string = '';
    IDNumber: string = '';
    Gender: string = '';
    DisplayName: string = '';
    Email: string = '';
    Password: string = '';
    EnableTwoFactorAuth: boolean = true;
}