// src/app/Model/superuser.ts
export class Superuser {
    id: string | undefined;
    title: string  = "";
    email: string = "";
    phoneNumber: string = "";
    first_Name: string = "";
    last_Name: string = "";
    iD_Number: string = "";
    hire_Date: Date | undefined;
    date_Created: Date | undefined;
    date_of_last_update: Date | undefined;
    username: string = "";
    twoFactorEnabled: boolean = true;
    gender: string = "";
    password: string = "";
}